refactor: use tasker's built-in scheduling for debounced billing
